Upgrade utility class to provide default merge and validate implementations
We are using Java introspection to get property fields.
We merge any subComponents by triggering merge on them when appropriate.
We merge fields according to whether they are null or not.
We have special handling for merging maps.
The default merge function assumes that the calling class fulfills the lombok @Builder contract

[rev. matthew.gordon]