Update all config classes in library to use new utility class and add standard tests
Added a new abstract skeleton implementation for ConfigurationComponent called SimpleComponent
Updated all config classes to extend SimpleComponent
Altered behaviour of DefaultLogin which was unobvious and asymmetrical: it previously claimed to be a wrapper around UsernameAndPassword, but did not do any "wrapping"; it only provided static methods. It expected to be parsed from json but never converted to json. Now it only has the static methods and is explicitly not a wrapper.
Added a standard abstract ConfigurationComponent test.
Wrote tests for all config classes in the library.
Shied away from finishing test for BCryptUsernameAndPassword as password behaviour is too opaque.

[rev. matthew.gordon]